+++ lib/util/mixins/tile_recognizer.dart
@@ -1,0 +1,69 @@
+import 'package:bonfire/base/game_component.dart';
+import 'package:bonfire/util/extensions/game_component_extensions.dart';
+
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 16/05/22
+
+mixin TileRecognizer on GameComponent {
+  /// Method that checks what type map tile is currently
+  String? tileTypeBelow() {
+    final list = tileTypeListBelow();
+    if (list.isNotEmpty) {
+      return list.first;
+    }
+    return null;
+  }
+
+  /// Method that checks what types map tile is currently
+  List<String> tileTypeListBelow() {
+    if (!hasGameRef) return [];
+    final map = gameRef.map;
+    if (map.getRendered().isNotEmpty) {
+      return map
+          .getRendered()
+          .where((element) {
+            return (element.overlaps(rectConsideringCollision) &&
+                (element.type?.isNotEmpty ?? false));
+          })
+          .map<String>((e) => e.type!)
+          .toList();
+    }
+    return [];
+  }
+
+  /// Method that checks what properties map tile is currently
+  Map<String, dynamic>? tilePropertiesBelow() {
+    final list = tilePropertiesListBelow();
+    if (list?.isNotEmpty == true) {
+      return list?.first;
+    }
+
+    return null;
+  }
+
+  /// Method that checks what properties list map tile is currently
+  List<Map<String, dynamic>>? tilePropertiesListBelow() {
+    if (!hasGameRef) return null;
+    final map = gameRef.map;
+    if (map.tiles.isNotEmpty) {
+      return map
+          .getRendered()
+          .where((element) {
+            return (element.overlaps(rectConsideringCollision) &&
+                (element.properties != null));
+          })
+          .map<Map<String, dynamic>>((e) => e.properties!)
+          .toList();
+    }
+    return null;
+  }
+}
